+++ Examples/Examples-iOS/IGListKitExamples/SectionControllers/HorizontalSectionController.swift
@@ -47,6 +47,8 @@ final class HorizontalSectionController: IGListSectionController, IGListSectionT
 
     func didSelectItem(at index: Int) {}
 
+    func didDeselectItem(at index: Int) {}
+  
     //MARK: IGListAdapterDataSource
 
     func objects(for listAdapter: IGListAdapter) -> [IGListDiffable] {